% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{beta_mom_estimator}
\alias{beta_mom_estimator}
\title{Beta Method-of-Moments Estimator}
\usage{
beta_mom_estimator(x)
}
\arguments{
\item{x}{A vector of beta-distributed values (between 0 and 1)}
}
\description{
A useful function for estimating the parameters for a beta distribution that
generated some given data. Returns estimated alpha and beta values
}
