% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{beta_plotter}
\alias{beta_plotter}
\title{Beta plotting function}
\usage{
beta_plotter(a, b, ...)
}
\arguments{
\item{a}{The alpha parameter of a beta distribution}

\item{b}{The beta parameter of a beta distribution}

\item{...}{Additional arguments to the plot function}
}
\description{
This function is simply used to plot a beta distribution, to give the user a
sense for what the simulated allele frequency distribution looks like.
}
