% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{bootstrap_comparison_test}
\alias{bootstrap_comparison_test}
\title{Define a new test for comparing the means of samples}
\usage{
bootstrap_comparison_test(haps_1, haps_2, exp_beta = "exp_beta", B = 1000)
}
\arguments{
\item{haps_1}{The output of `characterize_haplotypes`, or another haplotype calling tool. There is one line per haplotype. Must contain the column "beta".}

\item{haps_2}{The output of `characterize_haplotypes`, or another haplotype calling tool. There is one line per haplotype. Must contain the column "beta".}

\item{exp_beta}{The nuame of the column in "haps" that contains the expectation of beta. Defaults to "exp_beta"}

\item{B}{The number of bootstraps to run}
}
\description{
This function uses a very similar principal to the standard bootstrap test,
but it takes as input two sets of haplotypes, and compares if their epsilon
values are significantly different. This is useful when there is some bias
affecting both equally.
}
