% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{bootstrap_test}
\alias{bootstrap_test}
\title{Calculate a bootstrapped p-value}
\usage{
bootstrap_test(haps, exp_beta = "exp_beta", B = 1000)
}
\arguments{
\item{haps}{The output of `characterize_haplotypes`, or another haplotype calling tool. There is one line per haplotype. Must contain the column "beta" and "exp_beta"}

\item{exp_beta}{The nuame of the column in "haps" that contains the expectation of beta. Defaults to "exp_beta"}

\item{B}{The number of bootstraps to perform. Default: 1000}
}
\description{
This function takes the input data and returns a p-value for the null
hypothesis that haplotype configurations which influence coding variant
penetrance are not significantly enriched in the population. The p-value is
estimated through a bootstrapping procedure, and a confidence interval
is returned.
}
