% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{calculate_epsilon}
\alias{calculate_epsilon}
\title{Calculate epsilon}
\usage{
calculate_epsilon(dataset, beta_config, ...)
}
\arguments{
\item{dataset}{A set of haplotype counts}

\item{beta_config}{A beta configuration (use either beta_config_eqtl or beta_config_sqtl, which are included in this package)}

\item{...}{Additional arguments passed to `characterize_haplotypes`}
}
\description{
This function calculates the value of epsilon, dependent on the value of
beta and expected value of beta given each haplotype combination observed in
a set of haplotypes based on phased genetic data.
}
