% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{characterize_haplotypes}
\alias{characterize_haplotypes}
\title{Reshape data and classify the haplotypes according to the model}
\usage{
characterize_haplotypes(dataset, beta_config)
}
\arguments{
\item{dataset}{The input dataset. See sample data for format}

\item{beta_config}{The values assigned to beta for each haplotype possibility. This can be changed to tweak
the modified penetrance hypothesis. d

Internal function, not exported}
}
\description{
This function takes the haplotype configurations in the standard input format
and reshapes them for more intuitive processing. See the sample dataset
}
