% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{poison_binomial_test}
\alias{poison_binomial_test}
\title{Calculate a poisson-binomial p-value}
\usage{
poison_binomial_test(dataset, beta_config, ...)
}
\arguments{
\item{beta_config}{}
}
\description{
This function takes the input data and returns a p-value for the null
hypothesis that haplotype configurations which influence coding variant
penetrance are not significangtly enriched in the population. The p-value
is calculated analytically using the poisson-binomial distribuion, which
is the distribution over the sum of n independent and NOT identically
distributed bernoulli random variables.
}
