% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_haplotype_counts}
\alias{simulate_haplotype_counts}
\title{A function for generating fake data}
\usage{
simulate_haplotype_counts(n_indvs, n_genes, c_alpha = 4, c_beta = 3000,
  qtl_alpha = 1, qtl_beta = 1)
}
\arguments{
\item{n_indvs}{The number of individuals to simulate}

\item{n_genes}{The number of genes to simulate}

\item{c_alpha}{shape parameter for the csnp allele frequency. Default = 4}

\item{c_beta}{shape parameter beta for the csnp allele frequency. Default = 3000}

\item{qtl_alpha}{shape parameter alpha for the qtl allele frequency. Default = 1}

\item{qtl_beta}{shape paramteter beta for the qtl allele frequency. Default = 1}
}
\description{
This function can simulate some haplotype combinations based on allele
frequencies sampled from beta distributions. The qtl frequencies are assumed
to represent the higher expressed allele. qtl frequencies are drawn from a
beta distribution with a = b = 1
}
\examples{
simulate_haplotype_counts(n_indvs = 100, n_genes = 5)
}
