% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{bootstrap_test}
\alias{bootstrap_test}
\title{Calculate a bootstrapped p-value}
\usage{
bootstrap_test(haps, exp_beta = "exp_beta", af_weighted = FALSE, B = 1000, ...)
}
\arguments{
\item{haps}{The output of `characterize_haplotypes`, or another haplotype calling tool. There is one line per individual}

\item{B}{The number of bootstraps to perform. Default: 1000}

\item{beta_config}{The values assigned to beta for each haplotype possibility. This can be changed to tweak}
}
\description{
This function takes the input data and returns a p-value for the null
hypothesis that haplotype configurations which influence coding variant
penetrance are not significantly enriched in the population. The p-value is
estimated through a bootstrapping procedure, and a confidence interval
is returned.
}
