% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{bootstrap_test}
\alias{bootstrap_test}
\title{Calculate a bootstrapped p-value}
\usage{
bootstrap_test(dataset, beta_config, B = 1000)
}
\arguments{
\item{dataset}{The input dataset. See sample data for format}

\item{beta_config}{The values assigned to beta for each haplotype possibility. This can be changed to tweak}

\item{B}{The number of bootstraps to perform. Default: 1000}
}
\description{
This function takes the input data and returns a p-value for the null
hypothesis that haplotype configuations which influence coding variant
penetrance are not significantly enriched in the population.
}
