% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{calculate_epsilon}
\alias{calculate_epsilon}
\title{Calculate epsilon}
\usage{
calculate_epsilon(dataset, beta_config)
}
\arguments{
\item{dataset}{...}
}
\description{
This function calculates the value of epsilon, dependent on the value of
beta and expected value of beta given each haplotype combination observed in
a set of haplotypes based on phased genetic data.
}
