% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_test_functions.R
\name{characterize_haplotypes}
\alias{characterize_haplotypes}
\title{Reshape data and classify the haplotypes according to the model}
\usage{
characterize_haplotypes(dataset, beta_config,
  qtl_allele_frequency_column = "sqtl_af", gene_name_column = "gene")
}
\arguments{
\item{dataset}{A data.frame or tibble. The input dataset. See sample data for format}

\item{beta_config}{Vector. The values assigned to beta for each haplotype possibility. This can be changed to tweak the hypothesis.
Values are provided in the package as `beta_config_eqtl` and `beta_config_sqtl`, depending on the hypothesis
you are testing.}

\item{qtl_allele_frequency_column}{String. The name of the column in the input data with QTL allele frequencies. Default: `sqtl_af`
the modified penetrance hypothesis.}

\item{gene_name_column}{String. The name of the column in the input data with the gene name. Default: `gene`}
}
\description{
This function takes the haplotype configurations in the standard input format
and reshapes them for more intuitive processing. See the sample dataset
}
